Deployed, soon the client will come

+++ deploy.sh
@@ -14,7 +14,7 @@ where:
     -C  builds the stack
     -s  deploys the stack to the specified SSH_HOST, a docker swarm manager
     -S  deploys build macOS and Windows binaries to Steam
-    -m  clones the mongo database locally to ./.mongo
+    -m  clones the mongo database locally to ./.mongo, or to TARGET_MONGO_URL when set
 "
 deploy_steam=false
 deploy_stack=false
@@ -197,9 +197,15 @@ if [[ ${dump_mongo} == true ]]; then
   docker run --network=backend -i --rm mongo:3.6 bash -c "mkdir -pv ./out >/dev/null; mongodump  --excludeCollection=games --uri=mongodb://spellsource_mongo:27017/metastone >/dev/null; tar -czvf ./archive.tar.gz ./dump >/dev/null; cat archive.tar.gz" >"${archive_path}"
   mkdir -pv .mongo
   docker context use default >/dev/null
-  mongo_pid=$(docker run --rm -v="$(pwd)"/.mongo:/data/db -p="27017:27017" -d mongo:3.6)
+  if [[ -z ${TARGET_MONGO_URL} ]]; then
+    mongo_pid=$(docker run --rm -v="$(pwd)"/.mongo:/data/db -p="27017:27017" -d mongo:3.6)
+  else
+    mongo_pid=""
+  fi
   echo "mongod pid: ${mongo_pid}"
   tar -C "${dump_dir}" -xzvf "${archive_path}"
-  mongorestore --drop --uri=mongodb://localhost:27017/metastone "${dump_dir}"/dump
-  docker stop ${mongo_pid}
+  mongorestore --drop --uri="${TARGET_MONGO_URL:-mongodb://localhost:27017/metastone}" "${dump_dir}"/dump
+  if [[ -n ${mongo_pid} ]]; then
+    docker stop "${mongo_pid}"
+  fi
 fi
